Lab 5/requirements.txt
# ML
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.2.0
numpy==1.26.4
pandas==2.1.4
joblib==1.3.2

# Experiment Tracking
mlflow==2.12.1

# ML Metadata (lineage tracking)
ml-metadata==1.14.0

# Data Validation (TensorFlow Data Validation)
tensorflow-data-validation==1.14.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Web
flask==3.0.0
fastapi==0.109.0
uvicorn==0.27.0

# GCP
google-cloud-storage==2.14.0

# Airflow
apache-airflow-providers-postgres==5.10.0
psycopg2-binary==2.9.9
